error handling for drop day

+++ web-application-bp/app/weather/commands.py
@@ -8,5 +8,8 @@ import click
 def drop_day(day_id):
     """Delete specific day entry by id."""
     day = Day.query.filter_by(id=day_id).first()
-    db.session.delete(day)
-    db.session.commit()
+    if day:
+        db.session.delete(day)
+        db.session.commit()
+    else:
+        print(f"Day with id '{day_id}' does not exist")
